+++ build_autogenerated.yaml
@@ -2837,6 +2837,7 @@ libs:
   - include/grpcpp/impl/server_initializer.h
   - include/grpcpp/impl/service_type.h
   - include/grpcpp/impl/status.h
+  - include/grpcpp/impl/sync.h
   - include/grpcpp/resource_quota.h
   - include/grpcpp/security/auth_context.h
   - include/grpcpp/security/auth_metadata_processor.h
@@ -3262,6 +3263,7 @@ libs:
   - include/grpcpp/impl/server_initializer.h
   - include/grpcpp/impl/service_type.h
   - include/grpcpp/impl/status.h
+  - include/grpcpp/impl/sync.h
   - include/grpcpp/resource_quota.h
   - include/grpcpp/security/auth_context.h
   - include/grpcpp/security/auth_metadata_processor.h
