+++ gRPC-C++.podspec
@@ -163,6 +163,7 @@ Pod::Spec.new do |s|
                       'include/grpcpp/impl/server_initializer.h',
                       'include/grpcpp/impl/service_type.h',
                       'include/grpcpp/impl/status.h',
+                      'include/grpcpp/impl/sync.h',
                       'include/grpcpp/resource_quota.h',
                       'include/grpcpp/security/auth_context.h',
                       'include/grpcpp/security/auth_metadata_processor.h',
