+++ include/grpcpp/channel.h
@@ -25,8 +25,8 @@
 #include <grpcpp/completion_queue.h>
 #include <grpcpp/impl/call.h>
 #include <grpcpp/impl/channel_interface.h>
-#include <grpcpp/impl/codegen/sync.h>
 #include <grpcpp/impl/grpc_library.h>
+#include <grpcpp/impl/sync.h>
 #include <grpcpp/support/client_interceptor.h>
 #include <grpcpp/support/config.h>
