+++ include/grpcpp/client_context.h
@@ -41,10 +41,10 @@
 #include <grpc/impl/codegen/compression_types.h>
 #include <grpc/impl/codegen/propagation_bits.h>
 #include <grpcpp/impl/codegen/core_codegen_interface.h>
-#include <grpcpp/impl/codegen/sync.h>
 #include <grpcpp/impl/create_auth_context.h>
 #include <grpcpp/impl/metadata_map.h>
 #include <grpcpp/impl/rpc_method.h>
+#include <grpcpp/impl/sync.h>
 #include <grpcpp/security/auth_context.h>
 #include <grpcpp/support/client_interceptor.h>
 #include <grpcpp/support/config.h>
