+++ include/grpcpp/ext/call_metric_recorder.h
@@ -25,7 +25,7 @@
 #include "absl/strings/string_view.h"
 #include "absl/types/optional.h"
 
-#include <grpcpp/impl/codegen/sync.h>
+#include <grpcpp/impl/sync.h>
 #include <grpcpp/support/slice.h>
 
 namespace grpc_core {
