+++ src/cpp/common/tls_certificate_verifier.cc
@@ -28,8 +28,8 @@
 #include <grpc/support/alloc.h>
 #include <grpc/support/log.h>
 #include <grpc/support/string_util.h>
-#include <grpcpp/impl/codegen/sync.h>
 #include <grpcpp/impl/grpc_library.h>
+#include <grpcpp/impl/sync.h>
 #include <grpcpp/security/tls_certificate_verifier.h>
 #include <grpcpp/support/status.h>
 #include <grpcpp/support/string_ref.h>
