+++ src/cpp/server/health/default_health_check_service.h
@@ -29,8 +29,8 @@
 
 #include <grpcpp/grpcpp.h>
 #include <grpcpp/health_check_service_interface.h>
-#include <grpcpp/impl/codegen/sync.h>
 #include <grpcpp/impl/service_type.h>
+#include <grpcpp/impl/sync.h>
 #include <grpcpp/support/byte_buffer.h>
 #include <grpcpp/support/server_callback.h>
 #include <grpcpp/support/status.h>
