+++ src/cpp/server/orca/orca_service.cc
@@ -33,10 +33,10 @@
 #include <grpc/event_engine/event_engine.h>
 #include <grpc/support/log.h>
 #include <grpcpp/ext/orca_service.h>
-#include <grpcpp/impl/codegen/sync.h>
 #include <grpcpp/impl/rpc_method.h>
 #include <grpcpp/impl/rpc_service_method.h>
 #include <grpcpp/impl/server_callback_handlers.h>
+#include <grpcpp/impl/sync.h>
 #include <grpcpp/server_context.h>
 #include <grpcpp/support/byte_buffer.h>
 #include <grpcpp/support/server_callback.h>
