+++ tools/doxygen/Doxyfile.c++
@@ -1027,6 +1027,7 @@ include/grpcpp/impl/server_callback_handlers.h \
 include/grpcpp/impl/server_initializer.h \
 include/grpcpp/impl/service_type.h \
 include/grpcpp/impl/status.h \
+include/grpcpp/impl/sync.h \
 include/grpcpp/resource_quota.h \
 include/grpcpp/security/auth_context.h \
 include/grpcpp/security/auth_metadata_processor.h \
